fix(diet): 클라이언트 당류 절삭 제거 — drift·DTO 소수 정합
#362 로 서버는 당류를 소수로 내려주는데 클라이언트에 절삭 지점이 남아
데모 모드에서 8.5 가 8 로 잘렸다.

- drift `DietEntries.sugarG` IntColumn → RealColumn, schemaVersion 5 → 6
- `local_api_interceptor` 의 `.toInt()` → `.toDouble()`, 누산기도 double
- `RecognizedFood.sugarG`·`DietAnalysisResult.totalSugarG`·
  `NutritionDay.sugarG` 를 double 로 넓히고 파싱도 `.toDouble()`
- 시드 값 실수형 정리

마이그레이션은 테이블 재생성 없이 스키마 버전만 올린다. SQLite 의 INTEGER
친화도는 무손실일 때만 정수로 바꾸므로 8.5 는 REAL 로 남고, 읽을 때 drift 가
선언 타입으로 매핑한다. 기존 정수 값도 8 → 8.0 으로 손실 없이 읽힌다.
실험 단계 API(TableMigration)를 쓰지 않아도 되는 이유를 주석에 남겼다.

drift 라운드트립 테스트 3건으로 고정했다 — 8.5 저장·조회, v5 시절 정수 행이
double 로 읽히는지, 6.3+8.5=14.8 무절삭.

Closes #366

+++ frontend/flutter/lib/features/diet/domain/entities/diet_analysis.dart
@@ -12,7 +12,7 @@ class RecognizedFood {
   final String name;
   final int calories;
   final int sodiumMg;
-  final int sugarG;
+  final double sugarG;
   final String source; // "db"(공공 영양 DB 매핑) | "estimate"(LLM 추정)
 
   bool get isFromDb => source == 'db';
@@ -21,7 +21,7 @@ class RecognizedFood {
     name: json['name']! as String,
     calories: (json['calories'] as num?)?.toInt() ?? 0,
     sodiumMg: (json['sodium_mg'] as num?)?.toInt() ?? 0,
-    sugarG: (json['sugar_g'] as num?)?.toInt() ?? 0,
+    sugarG: (json['sugar_g'] as num?)?.toDouble() ?? 0,
     source: (json['source'] as String?) ?? 'estimate',
   );
 }
@@ -40,7 +40,7 @@ class DietAnalysisResult {
   final List<RecognizedFood> foods;
   final int totalCalories;
   final int totalSodiumMg;
-  final int totalSugarG;
+  final double totalSugarG;
   final String coachComment;
 
   factory DietAnalysisResult.fromResponse(Map<String, Object?> json) {
@@ -54,7 +54,7 @@ class DietAnalysisResult {
           .toList(),
       totalCalories: (analysis['total_calories'] as num?)?.toInt() ?? 0,
       totalSodiumMg: (analysis['total_sodium_mg'] as num?)?.toInt() ?? 0,
-      totalSugarG: (analysis['total_sugar_g'] as num?)?.toInt() ?? 0,
+      totalSugarG: (analysis['total_sugar_g'] as num?)?.toDouble() ?? 0,
       coachComment: (analysis['coach_comment'] as String?) ?? '',
     );
   }
